fix(canon): make runtime package resolution pluggable
