+++ docs/content/guide/troubleshooting.md
@@ -48,3 +48,12 @@ export default {
 Use `"strict"` to fail on invalid syntax, or `"quirks"` when a project relies on Vue accepting those
 tags as self-closing leaves. Valid void elements such as `<input />`, `<img />`, `<br />`, and
 `<meta />` do not need quirks.
+
+## Native Type Package Resolution
+
+`vize check` resolves Vue and Vite type packages from the checked project before it uses bundled
+fallbacks, so the project's own `vue`, `@vue/runtime-dom`, `@vue`, and `vite` versions drive the
+generated virtual project. For unusual package-manager layouts, set `VIZE_VUE_PACKAGE`,
+`VIZE_VUE_NAMESPACE_PACKAGE`, `VIZE_VUE_RUNTIME_DOM_PACKAGE`, or `VIZE_VITE_PACKAGE` to explicit
+package roots. `VIZE_RUNTIME_NODE_MODULES` can also point at one or more `node_modules` roots as a
+fallback search path.
